requirements.txt
# Ethos Research - Python Dependencies

# Ethos SDK (our official Python SDK)
ethos-py>=0.2.0

# Core
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Progress & Display
tqdm>=4.65.0
rich>=13.0.0

# Optional: Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
